requirements.txt
absl-py==0.8.1
altgraph==0.16.1
asgiref==3.2.3
astor==0.8.0
cachetools==3.1.1
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
Django==3.0
future==0.18.2
gast==0.2.2
google-auth==1.7.2
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.25.0
h5py==2.10.0
idna==2.8
image==1.5.27
imutils==0.5.3
joblib==0.14.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
numpy==1.17.3
oauthlib==3.1.0
opt-einsum==3.1.0
pefile==2019.4.18
Pillow==6.2.1
protobuf==3.11.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
pycrypto==2.6.1
PyInstaller==3.5
pyOpenSSL==19.1.0
PySocks==1.7.1
pytz==2019.3
pywin32==223
pywin32-ctypes==0.2.0
PyYAML==5.2
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.23.dev0
scipy==1.3.3
selenium==3.141.0
six==1.13.0
sqlparse==0.3.0
tensorboard==2.0.2
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
urllib3==1.25.7
Werkzeug==0.16.0
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.11.2
